Permit `b -> [0x333]`. Closes #4.
Now we just have the problem that `b <- [0x333]` and `b -> [0x333]` have
different immediate widths. This may need an issue of its own. The thing
to keep in mind is that those two instructions actually have two
different instruction formats, and that `b -> [0x333]` is actually just
shorthand for `b -> [a | a + 0x333]`.