Fix lib loading on Linux
Linux wasn't looking in current directory for .so files as Darwin was
for .dylib files, so now if the implname contains a path separator, we
treat it as a path instead of as a stem. In this case, spi.implstem must
also be set. implstem and implpath can also be used together to specify
multiple plugins to load from the same library, once we have instance
support.