Pass functions from plugin host into plugin
fatal_ and debug_ functions are now passed into plugins via a new plugin
entry point, tenyr_plugin_init(), if it exists. This resolves the
Windows build issues and the Linux runtime symbol resolution issues. The
current implementation of plugin_portable.h is a bit hacky in its use of
TENYR_PLUGIN and the way it affects common.h. This should be revisited.