Remove unsigned support from disassembler
This is the first of several commits, which will effectively remove
unsigned support from tenyr. The "unsigned" ops will remain, and behave
the same when operating on registers, but immediates will always be
sign-extended regardless of the operation being performed. If
sign-extension is not desired, the user will have to make do with an
eleven-bit immediate width.

The impetus for this change is largely to enable an optimisation for
branches, but it is also hoped that this simplification of the rules for
use of immediates will reduce confusion without reducing functionality.
It also reduces the complexity of implementation somewhat.