Remove output file in case of a problem
Because we can't remove a file by a connected stream, but rather only by
the filename we knew at the time we opened it, there is a race condition
in this code that could theoretically lead to data loss if you try
really hard. Therefore, the tenyr assembler is expressly prohibited from
being used in any high-reliability environment including but not limited
to life support, missions to Mars, and unmanned drone warfare.