Stop accounting for ambiguity in the disassembler
Previously, we tried to make assembly round-tripping work properly even
in the face of lossy disassembly (i.e. printing `A | B + 0` as `B`) but
now that we have the --verbose option to disassembly, this is no longer
required. Disassembly without --verbose will result in code with the
same behaviour, but may not reassemble into a bit-identical library ; if
this is desired, the --verbose option should be used.