Refactoring toward more intuitive disassembly
We would like to preserve the distinction between a number specified as
a bitstream (in some non-decimal base) and a number specified as an
integer (in decimal base) by correlating them with OP_BITWISE_OR and
OP_ADD when those ops are used as inert ops. This commit does not yet
effect that change, but it moves toward it.