Reduce the size of jzrel() and jnzrel()
Canonical branches now, since the introduction of immediates that are
sign-extended regardless of operation, can be two instructions long
instead of three.

It appears that the branch could be reduced to a single instruction if
P's read-value would return PC + 1. This is not unreasonable, but is a
fairly fundamental change to the tenyr architecture.