New jzrel() that works
Since 99cc85 jzrel() has been broken since Cond has not been on the
right side of &~. Nobody noticed because the only code that used jzrel()
was ex/sieve.tas.cpp, and its results are not easy to check.

Here is a jzrel() that works, and a simple testcase for it. It is one
instruction longer than jnzrel(), but it neither requires a temporary
register nor clobbers the condition register.