Support multiple plugin instances
For example :

    ./tsim -rplugin -p plugin.impl[0]=spi test/spi.texe -p spi.impl=spisd

or, more explicitly :

    ./tsim -rplugin --param=plugin.impl[0]=spi test/spi.texe \
        --param=spi.impl=./libspisd.so --param=spi.implstem=spisd

The `impl[N]` instances will be respected as long as they are contiguous
starting with 0 and can be loaded from the specified dynamic libraries.
`implstem` can be used to specify a stem within a library that may
contain multiple plugins, so a single library can serve up multiple
plugins. Currently the plugin architecture requires only one entry
point, `<implstem>_add_device`, conforming to the same interface used
for compiled-in plugins.