Make qsort comparison more fair
Use a function in qsort.c to make the comparison to the new
do_swap()-enabled code in qsort.tas.cpp more fair. Remove some
unnecessary stack traffic in qsort.tas.cpp. Although both codes are now
smaller, the size of the tenyr implementation relative to the x86-64
implementation is now greater than it was before the macro to function
conversion.