Support for FPGAlink programmer for Nexys3
Default to using flcli to program the Nexys3 instead of djtgcfg
(Digilent Adept), which has issues with segfaulting on Linux. Ensure
the normal user has permission to access the appropriate /dev/bus/usb
device files.

This workflow has not been completely vetted on Linux, as `impact`
currently catches a sigsegv and exits, but this it likely related to
the fact that my Linux installation is 13.4 and the bitstream was
generated by my (Windows) 14.4 installation.