Infer main RAM instead of using coregen
Under advice from MatthiasM@freenode, have XST infer the main memory
instead of using a coregen. This means our simmem becomes a real
BlockRAM module. Seems to work well in simulation, but not yet on
hardware. Other changes with timing are going on at the same time,
however, so it remains to be seen what is the root cause of the
hardware failure.