RESETVECTOR is now outside of device RAM
Finally, addresses < 0x1000 are left for devices. Adjusted prologue in
general to sit at `(1 << 13) - 1` because the previous `(1 << 14) - 1`
required the main RAM to be too large. If the main RAM is let grow large
enough to force the textram out of block RAM into FPGA fabric, the code
still runs but the VGA output shows nothing (although the MMRs are still
working).