Skeletal Core support for EIB
The EIB has not been hooked up, and the trap signal is currently tied
low. Some new wires have been introduced in Core whose names should be
subject to revision. Behaviour appears to be correct for non-trapping
conditions. Next step is to integrate the EIB (from eib.v) and test.