EIB now has interrupt vector table
Depending on the contents of the trampoline, it is not strictly
necessary for the EIB to handle the storage for the interrupt vector
table, but for convenience EIB will own it.

Reordered address checks to be correct (case inside if-tree instead of
if-tree inside case). There still may be a better way.