Only use @notation in memh output when needed
Don't use explicit addressing in memh output so that files can be loaded
into RAMs with different base addresses. For now, there's still no way
to specify a load offset to the assembler, so if there is a gap in the
source object (not really supported yet), the gap will be noted with
@offset that is relative to the beginning of the memh file (provoking
the same behaviour as before in Xilinx $readmemh -- balking at illegal
addresses).