Trying to reduce use of comparators / subtractors
Express RAMs as having configurable bottom addresses instead of using
explicit subtractors (may have the same effect : or will the synthesis
be smarter ?). Use masking instead of comparison to determine if an
address is in-range (forces ramwrap instances to be on power-of-two
boundaries but that's almost certainly what we want anyway).