Mend EIB mis-synthesis
Changing the IN_RANGE macro had caused the EIB to fail to synthesise
properly because it was receiving pre-sliced instead of whole addresses.

These changes revealed an apparent bug in Icarus, as well : macros in
different scopes, defined with the same name but differing number of
arguments, conflict and prevent compilation. This explains the separate
names for IN_RANGE, APROPOS and APPLIES.