Extend string and character parsing
Now \0 and \" are recognised in character escapes, and character escapes
are recognised inside strings. There are a few warts : \0 is swallowed
because it is returned as its own STRING, which gets strlen() called on
it. Pass back length instead of calling strlen() ever. Also, separate
STRING tokens are returned when one string should be built up in the
lexer in case it fails to parse eventually.