Mend failure with make_directive() since f29319d9
The commit f29319d9 introduced a sloppy error that passed a character
array instead of a struct strbuf into make_directive(). This was not
caught because there is no type-checking on the variadic arguments to
make_directive(). Now that `make check` works, we should be able to more
easily avoid such blunders.