Corrections and changes to video display hardware
Make text RAM start at base of video address space:

It's silly to reserve configuration register space at the bottom of the VRAM
address space. Instead, reserve a 4K block and put the registers at the top of
that.

There was apparently an RTL bug meaning the video RAM for the first 16 bytes
(supposed to be reserved for MMRs) was already addressing the text. This mends
that.

Mend garbage being written to display memory:

Since 1ad2eb2 which silenced a warning about bus widths mismatching, garbage
would be written to the display on basically every memory write, because
ramwrap.v was not enabled for only the area of memory devoted to the display :
it simply mapped the bottom 12 bits of the address to itself, and ignored the
bits above it.

bm_mults.texe now shows correct numbers on FPGA hardware, but exposes the fact
that current hardware does not know how to halt. The `illegal` instruction is
treated as a throw to a high (or negative) address, which overflows back to 0
and runs the code again. For now, halts should be replaced with infinite loops
while we consider how to implement software-initiated halt properly.

Mend bug with overlapping MMR and VRAM:

MMR reads were returning zero due to clashing values ; writes worked by
accident.

Now init_display clears the entire display.