Eliminate shift/reduce conflict that broke type2
From the introduction of type2 instructions it was supposed to be possible to
do `A <- 0x3 + B + C` but the use of `+` as the first op resulted in a
mis-parse as a type1 instruction. Hence `A <- 0x3 + B` was still a type1
instruction, expanded as `A <- A | 0x3 + B` ; now it is a type2 instruction,
expanded as `A <- 0x3 + B + A`.