Stop depending on `source ~/ise.sh`
Depend instead on a XILINX variable in the environment that points to
.../ISE_DS/ISE. Should work on 64-bit and 32-bit Linux.

Also make some cosmetic build changes and simplify the iMPACT process.

Clean up .xco generation

Even when there are no meaningful changes to the .xco file, it gets updated
with a new CRC and Date, which causes unnecessary rebuilds. When we build the
.xco, blank out these fields so they stay the same over rebuilds and checkouts.