Remove needless atomic moves
A badly-written Make rule could leave a file half-constructed or try to
consume it before it was finished being built (in a parallel make), which was
why we incrementally built some files to `$@.$$$$` and then moved to `$@` at
the end. These rules were not badly constructed, though, and testing by
inserting sleeps and killing the in-progress construction showed Make properly
disposing of the half-made file.