Refactoring reloc_expr
Simplifying grammar, at the cost of supporting examples like these :

    .word (@baz)        # do .word @baz instead
    .word ((@baz - .))  # do .word (@baz - .) instead

This may prove an undesirable restriction, but the existing reloc
grammar was hacky and we would rather have too many restrictions than
too few. When we had too few, we could also do this ;

	.word (@baz - . + 1)

but its semantics were the same as

	.word (@baz - (. + 1))

which is unacceptable.