Conway's Game of Life
Pick a seed lasting over 1200 generations.

Keep state as two bits in a 64*16 buffer instead of the obvious way (as
characters in two 80x40 buffers). This will permit us to fit in the
available SRAM for the current HDL implementation on Spartan6.

This currently requires a separate object (linked last) to contain the
buffer, in order not to relocate code too far away for a type{0,1,2}
instruction to access. We should be able to rewrite accesses to use
type3 for data, but we can't control what other codes are going to do,
so we should trap the too-large relocation offsets in the assembler.
This remains to be done.