Mending string handling
Because no token was returned to the parser until an entire string was
parsed, an unterminated string in a file could be ignored when it should
have been a syntax error. Now we return tokens for opening and closing
inverted commas, which increases complexity in the parser, but it also
improves diagnostic messages and removes the lexer-based limit on a
string's size.

Combined strbuf and cstr structures and updated code dependent on
parsing of strings and identifiers.