Stabilise VGA clock from DCM
The DCM_CLKGEN produces an apparently jittery clock when large divisors
are used, and to a certain extent even when the aspect ratio of M/D is
too high. A long-running DCM oscillator would also appear to decay,
exhibiting as a shaky output image.

Now use CLKFXDV to produce an exact quarter of the input frequency of
100MHz, after multiplication and division by 4. This produces what
appears to the most stable VGA pixel clock yet.