Guess version better
If we are not building from a Git repository, we won't be able to use
`git describe` to fetch our tag-based version information. In such a
case, try extracting a version tag by assuming we are being built from a
GitHub-generated release tarball. If that doesn't work, use the
modification date of the build directory, assuming that it has been
created by GitHub to reflect the time of the most recent commit
included. If all else fails, fall back to `unknown`.