Make `parse_opts_file` more robust
The `parse_opts_file` function provides configurability of `tsim` by allowing
options to be read from a file rather than from the command line. It attempts
to simulate the interpretive semantics of the shell in a superficial way. Now
it does a little better than before, by permitting a string of whitespace
separating two components of an argument (e.g. `-@ file` where previously
`-@file` was required).