Fix up bad `runwrap` assumption
For a long time (since 419e6f0a ?) we have assumed that `$(tas)` and
`$(tsim)` and `$(tld)` could be used both to build and to run the named
tools, but this worked only because we hadn't been testing `runwrap`.
Now that node.js support is somewhat functional again, the latent error
appears. Gingerly separate building from running with two sets of
variables (probably should rename the "run" set as well, or instead).