Run preamble earlier to get emscripten FS in place
The `os_preamble()` function (used so far only by emscripten) was
running too late to be effective for the `-@` option to `tsim`. Since we
no longer need a params object in `os_preamble()`, we no longer need to
wait until after args are parsed. Move the call to `os_preamble()`
higher in the tsim main function.