Add ASM_FORCE_DECIMAL_CONSTANTS flag
In normal mode, we disassemble to an unambiguous format that can always
be reassembled to the same instructions. Unfortunately, this means that
the common idiom

     P  <-  P + @+label

which assembles to a Type3 instructions for maximum reach, disassembles
to show a long hexadecimal constant even when the offset is usually a
smallish number that would be more readily understandable in decimal
format.

When ASM_FORCE_DECIMAL_CONSTANTS is set, we break our reversibility
guarantee in order to produce more human-pleasing disassembly output.
This flag is intentionally not exposed through the UI -- instead, it has
to be set through undocumented `--param` settings.