Fix bug in new-JIT implementation of `@`
There was a missing negation in the result of `@`, so a positive result
was 0x0000_0001 instead of 0xffff_ffff.

Introduce a mixer op test that has been demonstrated to catch this issue
and one hopes will catch others. There are only a few bits of entropy in
the output at present, due to the ad hoc nature of the test case
construction.