Revert "Address a latent bug in SDL devices"
This reverts commit 4e8f1237ffe2cb9015b34278d739d53a01caf09d.

The apparent bug with heap corruption appears to have been
misattributed, and explicitly destroying textures seems quite safe.

It is possible the memory corruption issues arose when using negative
addresses that are unmapped, which does indeed trigger undefined
behavior. This problem is being investigated separately. For now,
restore the SDL-related destruction behavior so as not to perpetuate an
apparently superstitious belief about SDL destruction, and also because
the removal of the destructions may cause a memory leak.