Switch to signed words in obj format
This could be considered a breaking change to object version 2, but we
are not at tenyr v1.0.0 yet, and nobody has objects that are large
enough to matter. "Thirty-one bits should be enough for anyone."

The "toolong" test case was updated to use the maximum signed 32-bit
size instead of the maximum unsigned 32-bit size.