Avoid creating symlinks inside symlinks
If `rsrc` already exists when we try to create it (for example if `make
-B` is specified), then a symlink could be created inside a symlink to a
directory -- nesting instead of overwriting -- due to the automatic
dereferencing of the target symlink.

GNU `ln` has `-T`, and BSD `ln` has `-F` (which is different), either of
which we might be able to use, but GNU `ln`'s `-F` appears to do
something different from BSD's, so it is not obvious how to write a
portable `ln` invocation. The safest bet is to remove something that
should be a symlink and replace it.