Remove forth implementation
The forth half-implementation, such as it was, was highly bitrotted, and
philosophically unsound (Forth words were being implemented in terms of
assembly tokens rather than in terms of text, with no end in sight).

A Forth implementation would be a delightful thing for tenyr to have,
and surely someday it will have one, but carrying around a partial
implementation that does not do anything useful except survive assembly
is not a good way to prepare for tenyr v1.0.