Drop support for web demo
The web demo is a desirable feature, and may return someday, but its
maintenance burden is not defensible for 1.0. On the other hand,
building for node.js (via asm.js or wasm) is a valuable capability that
will be maintained for 1.0.

Simplify use of preprocessor

Stop embedding .th and resource files

Drop tinypp

The only need for the tiny C preprocessor was to support the web demo,
which has been removed.

Drop (unused, non-working) SDL support under emscripten

Remove ui/web and disperse its contents

Stop using MODULARIZE