Rename `.utf32` -> `.chars`
The original aspiration to support UTF-32 as a natural consequence of
using 32 bits to represent a character has not been developed into an
actual implementation of UTF-32 in any meaningful way. It would be more
accurate to describe the assembler directive `.chars` as a way of
initializing memory using C-like character strings (including escapes)
in the current locale.

The current commit simply renames `.utf32` to `.chars` and keeps
functionality identical.

    git grep -il utf32 | xargs perl -i -pe 's/utf32/chars/g; s/UTF32/CHARS/g'