Consider it an error if counts are negative
In bfdb551f7de17d7b37cdc626614eded9533a7335, the object format replaced
UWord counts with SWord counts, but it did not start checking for
invalid negative numbers when parsing objects. This fact was obscured
because the negative tests involved failed as expected, but for the
wrong reason (fixed in a forthcoming, separate commit).