Remove buddy allocator
Since ebc65b83fdf278c3c417ffc231e175fe68277f1d, the buddy allocator no
longer compiled, but we did not notice because it was not being compiled
by default. As much as we would like to have a heap allocator, it does
not make sense to advertise one that does not even compile, and has no
tests. Perhaps this can be ressurrected at some point in the future.