Mark some unreachable code for lcov
Some fatal() conditions are actually unreachable, at least in the sense that
reaching them would require internal inconsistency in the codebase, and thus
it is not feasible to create a unit test to exercise them. Mark some of these
cases with LCOV_EXCL_* markers to avoid diluting true coverage metrics.