Convert some fprintf calls to fatal calls
Failure during simulation was sometimes returning zero due to a result variable
being overridden. Bailing out upon first error is more consistent with the
existing precedent.

This required an adjustment to the computation of PERIODS.mk, since now the
number of instructions executed will not be emitted if tsim bails early (such
as when the SDL plugin is not loaded).