Make Seg7 read on clock edge only
For some ancient reason, Seg7 (some of the oldest Verilog code in this project)
was readable without a clock edge. This will prevent good inference of
memories, and is unnecessary, so read out on a clock edge only.

(Reading out of Seg7 is not even a particularly necessary behavior, and could
be removed entirely at some point.)