Lengthen longest filename for loading into Icarus
The longest supported filename for loading into Icarus Verilog (via simtop.v)
was 100 characters. This was demonstrated to be insufficient when working in a
`git worktree` on macOS.

    % make -s --no-print-directory -C hw/icarus -f $PWD/build/arm64-apple-darwin20.4.0/PERIODS.mk -f Makefile run_code_modification VPATH=$PWD/test/op:$PWD/test/run BUILDDIR=$PWD/build/arm64-apple-darwin20.4.0 PLUSARGS_EXTRA=+DUMPENDSTATE
    WARNING: ../verilog/ram.v:33: $readmemh: Standard inconsistency, following 1364-2005.
    ERROR: ../verilog/ram.v:33: $readmemh: Unable to open default.memh for reading.
    Could not load file private/var/folders/x5/z29gc2pj4qq9wgkcm6_lw_hc0000gn/T/tmp.lmxHT9cm/test/run/code_modification.texe
    make: *** [run_code_modification] Error 1

Note that the initial slash in the "Could not load file" is missing -- the file
name was too long by one character.

We raise the limit to 4096 characters, which is a common maximum filename
length.