Strength the `@` test a bit
The amount of entropy tested by the `@` op-test is pretty low, and I have
observed on multiple occasions the test passing even when the hardware is not
doing its job. The current commit strengthens the randomization test a bit. An
even stronger test might ensure that a sole one or zero in a word of zeros or
ones is being extracted successfully.