Drop LCOV exclusions for lexer.l
The hacky workaround for an off-by-one `#line` numbering bug in flex is in the
way as I seek to simplify the build process. The bug still exists (one too many
blank lines are inserted in the resulting `lexer.c`), but since line coverage
is at 100% for `lexer.l` and should stay that way, nobody is liable to be
confused by this in the near term.