Remove an ambiguity-causing canonical sugaring
The instruction 0x80104000 disassembles as `A <- 0 + B + A` :

    $ tas -v -d -ftext - <<<0x80104000
     A  <-  0x00000000  +  B + A  # .word 0x80104000 ;            ; .addr 0x00000000

However, the canonical syntax sugaring resulted in `A <- 0 + B` :

    $ tas -d -ftext - <<<0x80104000
     A  <-  0  +  B               # .word 0x80104000 ;            ; .addr 0x00000000

When re-assembled, the latter instruction became 0x40010000, which sugars to `A <- B`.

    $ tas -d -ftext - <<<0x40010000
     A  <-  B                     # .word 0x40010000 ;            ; .addr 0x00000000
    $ tas -v -d -ftext - <<<0x40010000
     A  <-  A  |  0x00000000 + B  # .word 0x40010000 ;            ; .addr 0x00000000

This is equivalent in behavior, but tenyr explicitly holds out a guarantee that
every 32-bit word will survive disassembly and reassembly without changing its
bit pattern, and this was a failure of that guarantee.

This bug was discovered when expanding `test/pass_compile/ops.tas` to use
non-`A` registers and non-zero immediates. A subsequent commit to this one will
introduce these expansions.